fix: align safe block tag fallback and prevent blockSpan overflow
Make gRPC resolveBlockTag fall back to latest when finalized is zero
for the "safe" tag, matching the JSON-RPC shim behavior. Change
blockSpan return type from int to uint64 to prevent silent overflow
on large ranges that could bypass the max block range check and
cause a runtime panic in fetchBlockRange.

+++ architecture/evm/eth_query_helpers.go
@@ -518,11 +518,11 @@ func deepCopyQueryValue(value interface{}) interface{} {
 	}
 }
 
-func blockSpan(from uint64, to uint64) int {
+func blockSpan(from uint64, to uint64) uint64 {
 	if from >= to {
-		return int(from-to) + 1
+		return from - to + 1
 	}
-	return int(to-from) + 1
+	return to - from + 1
 }
 
 func blockMapFromRaw(raw json.RawMessage) (map[string]interface{}, error) {
